requirements.txt
pillow>=10.0.0
numpy>=1.24.0
pyyaml>=6.0.0
requests>=2.31.0
